change return value of variable nROE_ECLClusters from double to int

+++ analysis/tests/roe_variables.cc
@@ -209,11 +209,11 @@ namespace {
 
     var = Manager::Instance().getVariable("nROE_ECLClusters()");
     ASSERT_NE(var, nullptr);
-    EXPECT_FLOAT_EQ(std::get<double>(var->function(part)), 2.0);
+    EXPECT_FLOAT_EQ(std::get<int>(var->function(part)), 2);
 
     var = Manager::Instance().getVariable("nROE_ECLClusters(my_mask)");
     ASSERT_NE(var, nullptr);
-    EXPECT_FLOAT_EQ(std::get<double>(var->function(part)), 1.0);
+    EXPECT_FLOAT_EQ(std::get<int>(var->function(part)), 1);
 
     var = Manager::Instance().getVariable("nROE_NeutralECLClusters()");
     ASSERT_NE(var, nullptr);
